+++ server/component/style/radio/js/radio.js
@@ -1,11 +1,12 @@
 $(document).ready(() => {
-    check_locked_after_submit();
+    check_radio_locked_after_submit();
 });
 
-function check_locked_after_submit(){
+function check_radio_locked_after_submit(){
     $('.selfhelpRadio').each(function(){
+        console.log( );
         if($(this).data('locked_after_submit') && $('input[name="'+$(this).attr('name')+'"]:checked').val()){
-            $(this).prop('disabled', true);
+            $('input[name="'+$(this).attr('name')+'"]:not(:checked)').prop('disabled', true);
         }        
     })
 }